# generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Difficulty": {
        "enum": [
          "easy",
          "normal",
          "hard"
        ],
        "title": "Difficulty",
        "type": "string"
      },
...
      "EntityResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "image_attribution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageAttributionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "qid": {
            "title": "Qid",
            "type": "string"
          },
          "wikipedia_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wikipedia Url"
          }
        },
        "required": [
          "qid",
          "label",
          "description",
          "category",
          "entity_type",
          "image_path",
          "image_attribution",
          "wikipedia_url"
        ],
        "title": "EntityResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageAttributionResponse": {
        "additionalProperties": false,
        "description": "Complete attribution for a locally bundled Wikimedia Commons image.",
        "properties": {
          "attribution_text": {
            "title": "Attribution Text",
            "type": "string"
          },
          "context_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Label"
          },
          "creator": {
            "title": "Creator",
            "type": "string"
          },
          "derivative_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Derivative Url",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "license_id": {
            "enum": [
              "PUBLIC_DOMAIN",
              "CC0_1_0",
              "CC_BY_1_0",
              "CC_BY_2_0",
              "CC_BY_2_5",
              "CC_BY_3_0",
              "CC_BY_4_0",
              "CC_BY_SA_1_0",
              "CC_BY_SA_2_0",
              "CC_BY_SA_2_5",
              "CC_BY_SA_3_0",
              "CC_BY_SA_4_0"
            ],
            "title": "License Id",
            "type": "string"
          },
          "license_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "License Url",
            "type": "string"
          },
          "original_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Original Url",
            "type": "string"
          },
          "source_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "file_name",
          "original_url",
          "derivative_url",
          "source_url",
          "license_id",
          "creator",
          "license_url",
          "attribution_text"
        ],
        "title": "ImageAttributionResponse",
        "type": "object"
      },
...
      "RoomCloseReason": {
        "enum": [
          "not_enough_players"
        ],
        "title": "RoomCloseReason",
        "type": "string"
      },
      "RoomCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "round_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Round Id"
          }
        },
        "required": [
          "difficulty"
        ],
        "title": "RoomCreateRequest",
        "type": "object"
      },
      "RoomInviteResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "host_display_name": {
            "title": "Host Display Name",
            "type": "string"
          },
          "is_member": {
            "title": "Is Member",
            "type": "boolean"
          },
          "joinable": {
            "title": "Joinable",
            "type": "boolean"
          },
          "max_players": {
            "default": 4,
            "maximum": 4.0,
            "minimum": 4.0,
            "title": "Max Players",
            "type": "integer"
          },
          "player_count": {
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Player Count",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/RoomState"
          }
        },
        "required": [
          "code",
          "host_display_name",
          "state",
          "player_count",
          "is_member",
          "joinable"
        ],
        "title": "RoomInviteResponse",
        "type": "object"
      },
      "RoomParticipantResponse": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "finish_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Rank"
          },
          "guest_id": {
            "title": "Guest Id",
            "type": "string"
          },
          "hints_used": {
            "title": "Hints Used",
            "type": "integer"
          },
          "is_self": {
            "title": "Is Self",
            "type": "boolean"
          },
          "moves": {
            "title": "Moves",
            "type": "integer"
          },
          "progress_band": {
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Progress Band",
            "type": "integer"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "rematch_vote": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rematch Vote"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "guest_id",
          "display_name",
          "is_self",
          "active",
          "ready",
          "connected",
          "session_id",
          "moves",
          "hints_used",
          "progress_band",
          "finish_rank",
          "rematch_vote"
        ],
        "title": "RoomParticipantResponse",
        "type": "object"
      },
...
      "RoomRematchRequest": {
        "additionalProperties": false,
        "properties": {
          "accept": {
            "title": "Accept",
            "type": "boolean"
          }
        },
        "required": [
          "accept"
        ],
        "title": "RoomRematchRequest",
        "type": "object"
      },
      "RoomResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "close_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomCloseReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "countdown_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Countdown Ends At"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "grace_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grace Ends At"
          },
          "graph_version": {
            "title": "Graph Version",
            "type": "string"
          },
          "is_host": {
            "title": "Is Host",
            "type": "boolean"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/RoomParticipantResponse"
            },
            "title": "Participants",
            "type": "array"
          },
          "rematch_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rematch Ends At"
          },
          "round_id": {
            "title": "Round Id",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "start": {
            "$ref": "#/components/schemas/EntityResponse"
          },
          "state": {
            "$ref": "#/components/schemas/RoomState"
          },
          "target": {
            "$ref": "#/components/schemas/EntityResponse"
          }
        },
        "required": [
          "code",
          "state",
          "is_host",
          "graph_version",
          "round_id",
          "category",
          "difficulty",
          "start",
          "target",
          "participants",
          "sequence",
          "countdown_ends_at",
          "grace_ends_at",
          "rematch_ends_at",
          "close_reason"
        ],
        "title": "RoomResponse",
        "type": "object"
      },
      "RoomState": {
        "enum": [
          "lobby",
          "countdown",
          "racing",
          "grace_period",
          "finished",
          "closed"
        ],
        "title": "RoomState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Server-authoritative knowledge graph game API",
    "title": "Webwoven API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rooms": {
      "post": {
        "operationId": "create_room_api_v1_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{code}": {
      "get": {
        "operationId": "get_room_api_v1_rooms__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{code}/invite": {
      "get": {
        "operationId": "get_room_invite_api_v1_rooms__code__invite_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomInviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Room Invite",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{code}/join": {
      "post": {
        "operationId": "join_room_api_v1_rooms__code__join_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Join Room",
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/api/v1/rooms/{code}/rematch": {
      "post": {
        "operationId": "vote_room_rematch_api_v1_rooms__code__rematch_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomRematchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vote Room Rematch",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{code}/start": {
      "post": {
        "operationId": "start_room_api_v1_rooms__code__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Room",
        "tags": [
          "rooms"
        ]
      }
    },
...
  }
}